refactor(ui): move ProfileScreen into profile/ subpackage
Prep step for splitting ProfileScreen into nav + section
composables, mirroring the settings/ subpackage shape.
Pure rename + import-site fix.
